CTFs/Ethernaut/Telephone_DONE/pyproject.toml
[tool.poetry]
name = "telephone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
eth-ape = "*"
ape-foundry = "*"
ape-alchemy = "*"
ape-solidity = "*"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
